+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to the FolioBlocks project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.8] - 2026-02-18
+
+### Fixed 
+- Bug that prevented Play/Pause and Full Screen from being accessed when overlay was enabled on Filmstrip Gallery.
+
 ## [1.1.7] - 2026-02-17
 
 ### Changed